refactor: delete existing tag menu & get initItems from server (#25)
