+++ packages/mz-client/src/components/home/Home.tsx
@@ -6,69 +6,37 @@ import { useSearchParams } from 'next/navigation';
 import LinkCardList from '@/components/home/LinkCardList';
 import EmptyList from '../system/EmptyList';
 import { homeParameterStore } from '@/hooks/stores/HomeParameterStore';
-import { useGetFirstHomeItemsQuery } from '@/hooks/query/useGetHomeItemsQuery';
 import { shallow } from 'zustand/shallow';
 import Selectors from './Selectors';
 import LinkRowList from './LinkRowList';
-import { SortMode, View } from '@/types/db';
+import type { GetItemsResult, SortMode, View } from '@/types/db';
 import LoadMore from './LoadMore';
 import Button from '../system/Button';
-import { useQueryErrorResetBoundary } from '@tanstack/react-query';
-import SkeletonUI from '../system/SkeletonUI';
-import ErrorShower from '@/app/error';
 
-export default function Home() {
+export default function Home({
+  initItemsResult,
+}: {
+  initItemsResult: GetItemsResult;
+}) {
   const searchParams = useSearchParams();
   const startDate = searchParams.get('start');
   const endDate = searchParams.get('end');
   const [isMore, setIsMore] = useState(false);
-  const { reset } = useQueryErrorResetBoundary();
+  const { list, pageInfo, totalCount } = initItemsResult;
 
-  const { mode, view, tag, dateRange, setMode, setView, setDateRange } =
-    homeParameterStore(
-      (state) => ({
-        mode: state.mode,
-        view: state.view,
-        tag: state.tag,
-        dateRange: state.dateRange,
-        setMode: state.setMode,
-        setView: state.setView,
-        setDateRange: state.setDateRange,
-      }),
-      shallow
-    );
-  const { status, currentItems, error } = useGetFirstHomeItemsQuery({
-    mode,
-    tag,
-    dateRange,
-  });
+  const { mode, view, setMode, setView, setDateRange } = homeParameterStore(
+    (state) => ({
+      mode: state.mode,
+      view: state.view,
+      setMode: state.setMode,
+      setView: state.setView,
+      setDateRange: state.setDateRange,
+    }),
+    shallow
+  );
+
   const handleToggle = () => {
     setIsMore((prev) => !prev);
-  };
-
-  const getContent = () => {
-    if (status === 'loading') {
-      return <SkeletonUI />;
-    }
-
-    if (status === 'error') {
-      return <ErrorShower error={error as Error} reset={reset} />;
-    }
-
-    if (currentItems) {
-      const { list } = currentItems;
-      return list.length > 0 ? (
-        view === 'card' ? (
-          <LinkCardList items={list} />
-        ) : (
-          <LinkRowList items={list} />
-        )
-      ) : (
-        <EmptyList />
-      );
-    }
-
-    return null;
   };
 
   useEffect(() => {
@@ -90,17 +58,20 @@ export default function Home() {
   }, [startDate, endDate, mode]);
 
   // throw new Error();
-  const { list, totalCount, pageInfo } = currentItems ?? {
-    list: [],
-    totalCount: 0,
-    pageInfo: { hasNextPage: false, endCursor: null },
-  };
   const remainingItemCount = totalCount - list.length;
   return (
     <div className={styles.content}>
       <Selectors mode={mode} />
       {/* First item list section */}
-      {getContent()}
+      {list.length > 0 ? (
+        view === 'card' ? (
+          <LinkCardList items={list} />
+        ) : (
+          <LinkRowList items={list} />
+        )
+      ) : (
+        <EmptyList />
+      )}
       {/* Load more item list section */}
       {remainingItemCount > 0 && (
         <>
