+++ packages/mz-client/src/styles/SortModeSelector.module.scss
@@ -4,7 +4,6 @@
   display: flex;
   position: relative;
   gap: 16px;
-  margin-bottom: 24px;
 }
 
 .mode {
