+++ packages/mz-client/src/styles/ViewSelector.module.scss
@@ -1,35 +1,38 @@
 @import 'var';
 
 .menu {
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  display: inline-block;
   background: $gray01;
   height: 27px;
   border-radius: 27px;
+  margin-left: auto;
 
-  .menu_item {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 52px;
-    height: 26px;
-    border-radius: 27px;
-    cursor: pointer;
-    transition: all 0.3s ease;
+  li {
+    display: inline-block;
 
-    svg {
-      width: 16px;
-      height: 16px;
-      color: $gray2;
-    }
-
-    &.active {
-      border: 1px solid $primary;
-      background: white;
+    .menu_item {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      width: 52px;
+      height: 26px;
+      border-radius: 27px;
+      cursor: pointer;
+      transition: all 0.3s ease;
 
       svg {
-        color: $primary;
+        width: 16px;
+        height: 16px;
+        color: $gray2;
+      }
+
+      &.active {
+        border: 1px solid $primary;
+        background: white;
+
+        svg {
+          color: $primary;
+        }
       }
     }
   }
